refactor: Change repository structure.
