+++ lib/Lottotron.js
@@ -1,4 +1,4 @@
-const LottotronError = require('./lib/LottotronError')
+const LottotronError = require('./LottotronError')
 
 const ERROR_MSG = {
   MAX_NUMBER_IS_NOT_NUMBER: 'The input option "maxNumber" should be a number.',
