+++ package.json
@@ -2,7 +2,7 @@
   "name": "@spomni/lottotron",
   "version": "2.0.2",
   "description": "'Construct an object that returns random non-repeating integers'",
-  "main": "lottotron.js",
+  "main": "index.js",
   "directories": {
     "doc": "doc",
     "test": "test"
