+++ test/LottotronError.spec.js
